fix(redskilled): a command another package imports runs nothing on import
The #3546 branch delegated `dev statusline` to the daemon client by
importing `runStatusline` from `@reddb-io/redskilled/cli` — and cli.ts
ends in a self-invocation guard comparing `import.meta.url` against
`process.argv[1]`. Inside a single-file bundle every inlined module
shares the bundle's own URL, so the guard is ALWAYS true: the import
shipped a second, argv-reading CLI inside every dev binary. The
castle-mcp canary caught it as a pretty-printed host-state document
splashed across a stdout that must carry exactly one TOON report.

`runStatusline` moves to `statusline-command.ts`, a module that runs
nothing on import; cli.ts delegates and re-exports so its own surface
is unchanged, and the dev command imports the command module with a
comment naming the trap so the next consumer does not repeat it.
