+++ apps/redskilled/package.json
@@ -25,9 +25,10 @@
     "./statusline-payload": "./src/statusline-payload.ts",
     "./supervision": "./src/supervision.ts",
     "./worker-display": "./src/worker-display.ts",
-    "./worker-launch": "./src/worker-launch.ts"
+    "./worker-launch": "./src/worker-launch.ts",
+    "./statusline-command": "./src/statusline-command.ts"
   },
-  "description": "redskilled — the host-scoped execution daemon (ADR 0130): exactly one singleton per machine behind a unix socket, owning Worker processes across every project on that machine while each project's bundle keeps owning the work.",
+  "description": "redskilled \u2014 the host-scoped execution daemon (ADR 0130): exactly one singleton per machine behind a unix socket, owning Worker processes across every project on that machine while each project's bundle keeps owning the work.",
   "license": "Apache-2.0",
   "scripts": {
     "typecheck": "tsc -p tsconfig.json --noEmit",
